+++ pcap4j-core/src/main/java/org/pcap4j/core/PcapNetworkInterface.java
@@ -219,7 +219,7 @@ public final class PcapNetworkInterface {
       }
     }
 
-    return new PcapHandle(handle, true);
+    return new PcapHandle(handle);
   }
 
   @Override